# openapi.yaml
openapi: 3.0.3
info:
  title: no-sites-api
  description: Multitenant backend for health/medical sites (волосся.net, безсоння.net, etc.)
  version: 1.0.0
servers:
  - url: http://37.27.194.67
    description: Production (Hetzner)
  - url: http://localhost:3000
    description: Local dev
components:
  parameters:
    HostHeader:
      name: Host
      in: header
      required: true
      description: Site domain (e.g. volossia.net, bezsonnia.net)
      schema:
        type: string
        example: volossia.net
...
  schemas:
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          description: Response payload
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Record not found
...
    Session:
      type: object
      properties:
        session_token:
          type: string
          format: uuid
        expires_at:
          type: string
          format: date-time
          example: "2026-05-02T00:00:00.000Z"
        productIds:
          type: array
          items:
            type: string
            format: uuid
    Me:
      type: object
      properties:
        product_ids:
          type: array
          items:
            type: string
            format: uuid
          example: ["noalert_1", "noalert_2"]
...
paths:
  /auth/me:
    get:
      summary: Get products for current session
      tags: [Auth]
      description: |
        Validates the session token from the Authorization header and returns
        the list of product IDs the user has paid for on this site.
      parameters:
        - $ref: "#/components/parameters/HostHeader"
        - name: Authorization
          in: header
          required: true
          description: "Bearer <session_token>"
          schema:
            type: string
            example: "Bearer 550e8400-e29b-41d4-a716-446655440000"
      responses:
        "200":
          description: Session valid, product IDs returned
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Me"
        "401":
          description: Missing, invalid, or expired session token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Site not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/session:
    post:
      summary: Create session by email lookup
      tags: [Auth]
      description: |
        Looks up paid orders by email and site, creates a 30-day session token.
        Always returns 200 — if no orders found, productIds is empty (silent result).
        Response time is padded to a fixed minimum to prevent timing oracle attacks.
      parameters:
        - $ref: "#/components/parameters/HostHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
      responses:
        "200":
          description: Session created (productIds may be empty if no paid orders found)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Session"
        "400":
          description: Invalid or missing email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Site not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...